feat: move publish logic to write route, update publish to placeholder, and centralize address in layout
