+++ spec/00016_improve_menu_options_and_routes.txt
@@ -15,3 +15,4 @@ Implementation Notes:
 - Increment patch version in /package.json
 - Create a copy of this issue description and save to /spec using the name of this issue as the filename (with lower underscore case, starting with the 5 char padded issue number, e.g. 00016_issue_title.txt)
 - Unit test modules where applicable
+- Move the code in the publish route to the write route and put placeholder text in the publish route
