+++ src/routes/[address]/+layout.ts
@@ -1,0 +1,7 @@
+import type { LayoutLoad } from './$types';
+
+export const load: LayoutLoad = async ({ params }) => {
+  return {
+    address: params.address
+  };
+};
