+++ src/routes/[address]/+page.ts
@@ -2,7 +2,7 @@ import type { PageLoad } from './$types';
 
 export const ssr = false;
 
-export const load: PageLoad = async ({ params, fetch }) => {
+export const load: PageLoad = async ({ params, fetch, data }) => {
   const address = params.address;
   const url = `http://${address}/`;
   const res = await fetch(url, {
@@ -19,5 +19,5 @@ export const load: PageLoad = async ({ params, fetch }) => {
   } catch (e) {
     return { address, listing: [], error: 'Invalid JSON in listing' };
   }
-  return { address, listing };
+  return { ...data, listing };
 };
