--- src/routes/[address]/create/+page.ts
@@ -1,7 +1,0 @@
-import type { PageLoad } from './$types';
-
-export const load: PageLoad = async ({ params }) => {
-  return {
-    address: params.address
-  };
-};
