+++ src/routes/[address]/publish/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import CreatePage from '../create/+page.svelte';
   export let data: { address: string };
   const address = data?.address ?? '';
   $: pageTitle = `${address} - Publish - IMIM 2.0`;
@@ -9,4 +8,8 @@
   <title>{pageTitle}</title>
 </svelte:head>
 
-<CreatePage {data} />
+<div class="space-y-4">
+  <h1 class="text-2xl font-semibold">Publish</h1>
+  <p class="text-sky-700">Once you have finished writing your article, you can publish it to the network from this page.</p>
+  <p class="text-sky-600 text-sm">Note: This is a placeholder for the future publishing workflow.</p>
+</div>
